rcu: Use data_race() for RCU CPU stall-warning prints
Although the accesses used to determine whether or not a stall should
be printed are an integral part of the concurrency algorithm governing
use of the corresponding variables, the values that are simply printed
are ancillary.  As such, it is best to use data_race() for these accesses
in order to provide the greatest latitude in the use of KCSAN for the
other accesses that are an integral part of the algorithm.  This commit
therefore changes the relevant uses of READ_ONCE() to data_race().

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>
Change-Id: Ibdfbb80355dcefaf8f172e7abdbc271e6b7f7a51
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
(cherry picked from commit 324d51ef3c42f16b45a7fac9ee766c8217e82348)
Signed-off-by: TogoFire <togofire@mailfence.com>
Signed-off-by: RizkyMaulana58 <carlotta.montelli.0201@gmail.com>