scheduler - add entropy sampling hook
The scheduler can be used as a source of entropy. This requires the
presence of a hook that invokes the entropy source implementation.

When the scheduler-based entropy source is not compiled, the hook is
folded into a noop which does not affect the scheduler in any way.

Signed-off-by: Stephan Mueller <smueller@chronox.de>
Change-Id: I3afc1b7103ca8292c6bf6b8d356caa4a23ce362e
(cherry picked from commit 76d116c19dc1f72675c485d7b661edae0c95492c)
Signed-off-by: TogoFire <togofire@mailfence.com>
Signed-off-by: RizkyMaulana58 <carlotta.montelli.0201@gmail.com>