LRNG - add random.c entropy source support
The random.c implementation can be used as an entropy source by the
LRNG. This support can be enabled at compile time.

The entropy rate can be set at compile time which is only applied:

- once the random.c considers itself fully seeded, and

- the kernel does not operate in FIPS mode (i.e. fips=1 is not set at
  the kernel command line)

If one of these properties is not set, the ES will obtain data, but will
credit it with zero bits of entropy. For the first bullet, it is clear
why it will have zero bits of entropy. But for the second property, this
is set because the random.c is not operating SP800-90B compliant and
thus must be treated to not deliver any entropy in FIPS mode.

Signed-off-by: Stephan Mueller <smueller@chronox.de>
Change-Id: I135e4a4f2f47f0b8c57dd0bf8ffb13f45721976e
(cherry picked from commit 6b9d69ec952411be8cf0d64e624397dd77390d60)
Signed-off-by: TogoFire <togofire@mailfence.com>
Signed-off-by: RizkyMaulana58 <carlotta.montelli.0201@gmail.com>