techpack/dataipa: affine ipa irqs to LITTLE CPUs
Some of the ipa irqs are run on CPU0 only and a few are run on CPU6.
Instead of utilizing the big clusters, affine these
completely to little cluster.
Little cluster can easily handle these interrupts and manage them
in a power efficient manner rather than handing these to the big
clusters.

[mornye1517: Adapt to sm6375]

Change-Id: Ie4b69cf3c927aee6275dfafa3e81a01783781a15
Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>
Signed-off-by: Mornye <mornye150190@gmail.com>